latest versionnnn

+++ src/main/java/com/mcore/managers/TeleportMath.java
@@ -1,8 +1,3 @@
-//
-// Source code recreated from a .class file by IntelliJ IDEA
-// (powered by FernFlower decompiler)
-//
-
 package com.mcore.utils;
 
 import org.bukkit.Location;
@@ -10,9 +5,9 @@ import org.bukkit.util.Vector;
 
 public class TeleportMath {
     public static Location[] getFacingLocs(Location center, double distance) {
-        double half = distance / (double)2.0F;
-        Location loc1 = center.clone().add(half, (double)0.0F, (double)0.0F);
-        Location loc2 = center.clone().subtract(half, (double)0.0F, (double)0.0F);
+        double half = distance / 2.0;
+        Location loc1 = center.clone().add(half, 0, 0);
+        Location loc2 = center.clone().subtract(half, 0, 0);
         lookAt(loc1, loc2);
         lookAt(loc2, loc1);
         return new Location[]{loc1, loc2};
